requirements.txt
fastapi
uvicorn
crewai
langchain-google-genai
pypdf
python-dotenv
langchain-community